requirements.txt
# Django and Core
Django>=5.0,<6.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0

# Django Extensions
django-crispy-forms>=2.1
crispy-tailwind>=1.0.0
django-widget-tweaks>=1.5.0

# AI/ML Libraries
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0

# Data Processing
openpyxl>=3.1.0
reportlab>=4.0.0

# Development
django-debug-toolbar>=4.2.0

# Security
django-axes>=6.1.0

# ID Card OCR (free, local, supports Arabic + English)
easyocr>=1.7.0
